WordPress-XML-RPC-Password-Brute-Forcer
This Bash script is designed to perform brute-force attacks on a WordPress-based website. Its main function is to attempt to guess the password of a provided user through a password list file named "rockyou.txt." When the script is executed, it prompts the user to enter a WordPress username and then iterates through the passwords in the list file, generating XMLRPC requests that attempt to authenticate to the website using the username and each password from the list. If it finds a valid password, it displays it in the output.