maliciousSiteRedirection.js
When this script is maliciously injected into a form through a Cross-Site Scripting (XSS) vulnerability, it triggers an unauthorized redirection action. This occurs when the script exploits the lack of data validation and escaping in the web application, allowing the attacker to execute arbitrary JavaScript code in the user's browser.