noSQLI.py
This Python3 script performs a brute-force attack using a NoSQL Injection technique to decrypt the password of the "admin" user on a local web system utilizing a NoSQL database. The script attempts to guess each character of the password one by one, employing a brute-force technique and an HTTP POST request with a regular expression filter on the password in JSON format. When it finds a correct character, it appends it to the password and continues the process until the complete password for the "admin" user is decrypted.

IMPORTANT!:

- Make sure to modify the login_url variable with the address of the website you want. And the username of whom you want to have their passwords in the post_data variable