latexInjection.sh
This Bash script performs a series of HTTP requests to a local server in search of LaTeX injections in a specific file (specified as a command line argument). It then downloads the files generated by the server, converts them from PDF to text, reads the first line of each downloaded text file, and prints them to the console.

IMPORTANT:
To customize the script, you need to change the server URL (main_url) to the address of the server you want to use. You can also modify the loop iteration count if you want a different number of download attempts and the file path you want to view.