squidProxyPortDiscovery.py
This Python script performs a common port discovery on a local server (127.0.0.1) through HTTP requests to those ports. It utilizes a Squid proxy (192.168.71.131:3128) for the requests. If a port responds with a code other than 503, it indicates that the port is open, printing an informative message. The script handles Ctrl+C interruption for a controlled exit.

IMPORTANT:
To customise the script, you must change the server URL (main_url) to the server address you wish to use. You also have to modify the ip of the Squid_Proxy but without removing the port which is always the same.