shellShock.py
This Python script performs a Shellshock attack through a vulnerability in the Bash interpreter. It sends an HTTP request to a CGI service at http://127.0.0.1/cgi-bin/status with a manipulated User-Agent header to execute remote commands on the target machine. In addition, it uses a Squid proxy (192.168.71.131:3128) for requests. The script starts a thread to execute the attack and waits for a connection on port 443 to interact with the remote shell if the exploit succeeds. If the connection is not established within 20 seconds, it displays a failure message and exits the script.

IMPORTANT:
To customise the script, you must change the server URL (main_url) to the server address you wish to use. You also have to modify the ip of the Squid_Proxy but without removing the port which is always the same.