subtagFiltering.py
This XPath injection script is used to perform a brute-force attack on a website that is vulnerable to XPath injections. The goal is to extract information from the underlying XML by forcing evaluations of XPath expressions. The script iterates through the positions and characters of the XPath expressions, making HTTP requests to determine if the condition is true. If it is, the script collects the corresponding character at that position. The final result is the disclosure of the content of the main XML element.

IMPORTANT:
You need to modify the main_url and the error code you obtain in the variable r.text.